requirements.txt
Flask==3.0.0
python-docx==1.1.0
pathlib
Werkzeug==2.3.7
functions-framework==3.5.0
gunicorn==21.2.0 